Inefficient even-grid SIFT implementation drafts.
Attempted to implement evenly-grid + SIFT descriptor, as some paper demonstrated that it yielded better result than DoG blob detection.

However, the large number of the images makes the implemetation not efficient enough for the short amount of time for this assignment.

Given the limited amount of time available for this assignment, there is little hope to improve the implementation to support the datasets we have.

Hence, just commit it for record.